Remove allForks namespace
