+++ packages/types/package.json
@@ -17,9 +17,6 @@
     ".": {
       "import": "./lib/index.js"
     },
-    "./allForks": {
-      "import": "./lib/allForks/index.js"
-    },
     "./altair": {
       "import": "./lib/altair/index.js"
     },
